OCDAV: fix returned timestamp format
RFC1123 actually recommend the use of "UT" as zimezone instead of "UTC".
The way to do it with go is using the special "RFC1123Z" time.

Fixes issue with timestamp warnings on the client and makes the format
compliant with the old implementation.

See https://github.com/golang/go/issues/13781#issuecomment-168235353