Propagate upload mtime to the storage
The value of X-OC-Mtime is now passed to InitiateFileUpload using an OpaqueEntry.

The storage FS interface's InitiateUpload was updated to include a new
metadata argument, and adjusted all known implementors.

This new metadata argument is used in the storage provider to pass in the
mtime to the storage implementations.

The ownCloud storage implementation was adjusted to read the mtime from the
metadata and set it using Chtimes. That logic was extracted in a new
private method to avoid gymnastics with path handling and also avoid the
implicit propagation that exists in SetArbitraryMetadata.

The other implementors like EOS and Local are currently ignoring the
mtime value and need to be adjusted later on.