downgrade grpc to 1.26.0
In ocis we currently face the problem that some of our dependencies rely on
grpc 1.26.0. With the upgraded grpc dependency in reva we were unable to
build ocis. And if we upgrade grpc in ocis the build failed because of
the other dependency.

I hoped we could downgrade grpc in reva until we have found a better
long term solution.

Here are some examples of our attempts to resolve this without this
downgrade. https://github.com/owncloud/ocis/pull/328#issuecomment-647455011