Fix OCS public link share update values logic
Only update a value if it was actually set in the form.

Added support for clearing expiration date and password by
distinguishing between an empty string and whether the value is actually
set in the request.

Refactored public permission parsing to use permissionFromRequest for
both creation and update of public shares.