fix public storage stat
When doing a stat request in the public storage we need to differentiate between folder and file shares. If we want to stat a file in folder shares we need to concatenate the originalPath and the relative path because the originalPath is only the path to the share root. That means only up to the shared folder and the relative path contains the file path from the share root.
The originalPath in file share is already pointing to the shared file so
we don't need to append the relative path there.