Switch LDAP test to use entryUUID as the user's unique id
The entryUUID attribute is generated by the LDAP server. By using this
me can make sure that the the users get different IDs with every
test. So that can avoid setting DELETE_USER_DATA_CMD to delete
the USER data after each test in the long run.

Also this seems to uncover some issue in the WebDAV code which still
assumes that the userid always match the username.