Check public share child items in the auth scope handler
This solution is by far not pretty. But this is the smallest change I
could find to support apps in public shares without completely rewriting
the scopes code.
If we want to use any app in a public share the app needs an
authenticated context to access the files in the share. This is working
fine since we can authenticate using the share token. But the resulting
JWT contains a scope refering to the share resource. That means for a
folder share it contains the resource ID of the shared folder. But if
the app wants to access a specific file in the share it will request
that file by ID. Because of that the old way of the public scope checker
doesn't work because it would check if we are accessing the ID of the
shared folder and not if we are accessing any file in that folder.
Because of this problem I added two additional Stat requests so that I
can check if the requested file is inside the shared
folder.