Allow to run user_provider integration test with LDAP driver (#2584)
* Allow to run user_provider integration test with LDAP driver

The test are skipped by default for now, as they require an LDAP Server
to be setup. To run you can e.g. use OpenLDAP in docker use e.g.:

```
docker run --name openldap \
            --volume ./tests/integration/grpc/fixtures/ldap/:/container/service/slapd/assets/config/bootstrap/ldif/custom \
            -e LDAP_DOMAIN="owncloud.com" \
            -e LDAP_TLS_VERIFY_CLIENT=never \
            -e LDAP_ORGANISATION=ownCloud \
            -e LDAP_ADMIN_PASSWORD=admin \
            -e LDAP_REMOVE_CONFIG_AFTER_SETUP=false \
            osixia/openldap
```

run the test suite with:

```
cd tests/integration/
RUN_LDAP_TESTS=1 ~/go/bin/ginkgo [--focus "ldap"] ./...
```

* LDAP GetUser/GetGroup take IDP attribute into account

If GetUser/GetGroup is called with the IDP attribute set to a non-empty
string we only return results if the IDP setting configured for the
user/groupprovider matches the IDP from the request. This aligns the LDAP
providers' behaviour with what the demo and json providers implement.

* Address ginkgo v2 deprecation warnings

CurrentGinkgoTestDescription() is deprecated in V2 and should be
replaced with CurrentSpecReport().