LDAP: rework configuration and move to a common module
This reworks the LDAP configuration of the user- and groups-providers.
As the LDAP groups provider needs to be aware of users (e.g. for
looking up groups members) and the LDAP users provider needs to	be aware
of groups (e.g. for looking up groupmembership of a user) this commit
basically unifies the configuration of both providers.

Additionally the LDAP configuration is reworked to no longer rely on
templating LDAP filters in the configuration with can be error prone and
confusing.

This commit also move all code executing LDAP queries into the common
utlis/ldap module for being able to share more code between both
providers.