LDAP: rework configuration and move to a common module (#2708)
* ldap: Use a long lived LDAP connection where possible

Instead of open a new connection with every single requests. We now use
a long-lived connection per service with a simple mechanism to
automatically reconnect in case of failures.

Fixes #2122

* LDAP: rework configuration and move to a common module

This reworks the LDAP configuration of the user- and groups-providers.
As the LDAP groups provider needs to be aware of users (e.g. for
looking up groups members) and the LDAP users provider needs to	be aware
of groups (e.g. for looking up groupmembership of a user) this commit
basically unifies the configuration of both providers.

Additionally the LDAP configuration is reworked to no longer rely on
templating LDAP filters in the configuration with can be error prone and
confusing.

This commit also move all code executing LDAP queries into the common
utlis/ldap module for being able to share more code between both
providers.

* LDAP: Allow to work the standard LDAP groups

Depending on the configured group_objectclass we can no resolve group
members by DN as well.

Closes: 2124

* Add LDAP errors to debug log

* Move retry logic into a separate function

* Rework LDAP authprovider to use shared LDAP code

* Fix getUser call in owncloudsql storage provider

GetUser() was used to lookup users by name or ID, however it is only
supposed to work for ID based lookups. If the ID based lookup fails
fallback to an explicit GetUserByClaim fall for trying a name based
lookup.