Add separate option to configure tls mode for permissions client
This is a (temporary) workaround for allowing to separately configure the
tls mode for the permissions client used by the decomposedfs storage. This
is needed for being able to use the ocis settings service as the permission
service untils ocis' go-micro based services are adapted for providing TLS
as well.