Fix nested resource check for publicly shared space root
The 'Path' returned as part of the ResourceInfo from a Stat call might
be relative. So we need to Use 'GetPath()' to lookup the full path of
the parent. Otherwise the simple 'HasPrefix' check for whether a file a
ancestor might fail.

Co-Authored-By: Christian Richter <crichter@owncloud.com>
(cherry picked from commit 56d2be7011cf61ee9c491991eb75bdde70f56c71)