fix(decomposedfs): available space calculation
Move the calculation of the available size to the blobstore interface.
We were just returning the available size of the local disk, which is
wrong when using the S3 blobstore.

Also, as the S3 blobstore doesn't have a concept of available size, we
don't return a 'remaining' for S3 backed spaces with unlimited quota.

Fixes: https://github.com/owncloud/ocis/issues/9245