fix(ocm): Fix IDP value in federated user ids
Up to now the IDP for federated users id was set to the domain
of the user's idp. We need to set it to the federation provider's
domain name so that we can check it against the provider domain
as configured in the provider info (as e.g. read from providers.json)