New "jsoncs3" backend for app token storage
This backend stores the apptoken in a JSON file per user using the
metabackend. By default passwords are generated using the "diceware"
algorithm (based on https://xkcd.com/936/). It can be switched to
generate "normal" random password. Passwords are hashed using the
argon2id algorithm.
Note: There is no form of caching for now. I.e. the json file is loaded
upon every single authentication request so there is lot of room for
improvement here. To limit the amount of write operations the Utime
(time of last usage of the password) is however only updated with a
granularity of 5 minutes.

https://github.com/opencloud-eu/opencloud/issues/207