Fix stale file metadata cache entries
When moving a file to the trashbin we need to purge the file metadata
cache in the metadata backend (AFTER moving the file to the trash so the
metadata isn't removed from the file on disk). Otherwise the data will
still be read from the cache e.g. from ReadNode(), masking the fact that
the node is no longer available.