fix infinite loop with mtime
When you fetch a storage space, a Stat call happens to the root of the storage space to get the space's mtime.
However, to allow project admins to fetch all shares, we would need to set the `.Space` property of a resource info in a Stat, which could lead to an infinite loop. Therefore, this commit only does the fetching of a space when a special value in the opaques is set.

To do this efficiently, a space info cache was also added to the StorageProvider