Show shares to project admins (#5301)
* Show shares to project admins

Some modifications to the ways shares are listed to allow project admins to view shares

* fix unit tests

* add user filter to SharedByMe

This auto-injected filter was removed in the plugins by https://github.com/cernbox/reva-plugins/pull/51

* Making cache generic

* fix infinite loop with mtime

When you fetch a storage space, a Stat call happens to the root of the storage space to get the space's mtime.
However, to allow project admins to fetch all shares, we would need to set the `.Space` property of a resource info in a Stat, which could lead to an infinite loop. Therefore, this commit only does the fetching of a space when a special value in the opaques is set.

To do this efficiently, a space info cache was also added to the StorageProvider